Rendu navbar fixe et ajouté bon icônes à quelques pages
