+++ src/assets/scss/_layout.scss
@@ -1,2 +1,2 @@
 $margin-step: 16px;
-
+$navbar-height: 72px;
