+++ src/components/oeuvre/page-fichiers.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { Translation } from "react-i18next";
 import { PageAssistant } from '../canevas/page-assistant';
 import { ChampDate } from "../formulaires/champ-date";
-import PlusCircleOrange from '../../assets/svg/icons/plus-circle-orange.svg';
-import PlusCircleGreen from '../../assets/svg/icons/plus-circle-green.svg';
+import FileCircleOrange from '../../assets/svg/icons/file-circle-orange.svg';
+import FileCircleGreen from '../../assets/svg/icons/file-circle-green.svg';
 
 export default class PageFichiers extends React.Component {
 
     icon() {
-        return this.props.pochette ? PlusCircleOrange : PlusCircleGreen;
+        return this.props.pochette ? FileCircleOrange : FileCircleGreen;
     }
 
     render() {
