+++ src/components/oeuvre/page-influences.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { Translation } from "react-i18next";
 import { PageAssistant } from '../canevas/page-assistant';
 import { ChampDate } from "../formulaires/champ-date";
-import MusicOrange from '../../assets/svg/icons/music-orange.svg';
-import MusicGreen from '../../assets/svg/icons/music-green.svg';
+import MusicCircleOrange from '../../assets/svg/icons/music-circle-orange.svg';
+import MusicCircleGreen from '../../assets/svg/icons/music-circle-green.svg';
 
 export default class PageInfluences extends React.Component {
 
     icon() {
-        return this.props.pochette ? MusicOrange : MusicGreen;
+        return this.props.pochette ? MusicCircleOrange : MusicCircleGreen;
     }
 
     render() {
